Remove more files, change types, add src directory

+++ src/Planet.java
@@ -1,13 +1,13 @@
 
 public class Planet {
 	
-	public float G;
+	public double G;
 	
-	Planet(float g) {
+	Planet(double g) {
 		this.G = g;
 	}
 	
-	public float t = 0;
+	public double t = 0;
 	public int x = 1000; // horizontal canvas size 
 	public int y = 450; // vertical canvas size
